Adding loading styles file
