+++ frontend/src/styles/loading-styles.ts
@@ -1,0 +1,9 @@
+
+export const LOADING_SCREEN_TEXT_STYLE = {
+  font: "20px monospace",
+  color: "#000000",
+};
+export const LOADING_SCREEN_PERCENT_TEXT_STYLE = {
+  font: "18px monospace",
+  color: "#a6af48",
+};
